requirements.txt
Flask==3.1.3
gunicorn==21.2.0

PyMuPDF==1.23.26
pillow==10.3.0
python-docx==1.1.0
pdf2image==1.17.0
pytesseract==0.3.10
numpy==1.26.4
openpyxl==3.1.2

watchdog==3.0.0

google-cloud-vision==3.4.4
google-api-core==2.17.1
google-auth==2.29.0

ollama==0.1.8
requests==2.31.0

